requirements.txt
colorama==0.4.6
pillow==11.1.0
tqdm==4.67.1
